+++ observatory-platform/observatory/platform/dags/vm_create.py
@@ -17,7 +17,7 @@
 import pendulum
 from airflow import DAG
 from airflow.operators.python import PythonOperator, ShortCircuitOperator
-from observatory.platform.tasks.terraform import TerraformTasks
+from observatory.platform.workflows.terraform import TerraformTasks
 
 default_args = {"owner": "airflow", "start_date": pendulum.datetime(2020, 7, 1)}
